00:52 2/7/2017

+++ database/migrations/2017_06_14_012046_add_rate_to_worker.php
@@ -15,7 +15,7 @@ class AddRateToWorker extends Migration
     {
         Schema::table('workers', function (Blueprint $table) {
             //
-            $table->unsignedInteger('rate')->after('wage')->default(0);
+            $table->float('rate')->after('wage')->default(0);
         });
     }
